+++ src/Commands/ImportDump.php
@@ -13,6 +13,7 @@ use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Str;
 use function Laravel\Prompts\confirm;
 use function Laravel\Prompts\error;
 use function Laravel\Prompts\info;
@@ -40,9 +41,7 @@ class ImportDump extends Command
                 {--c|connection= : The configured database-connection in Laravel\'s config/database.php. }
                 {--allow-production : Enable importing SQL dumps on a production system. }
                 {--force : Skips confirmation prompts. Requires the file, remote or latest option. }
-                {--i|ignore-connection-filter : Ignores filter of dumps to defined connections. }
                 {--r|remote : Pull a fresh dump from the remote server as configured in the .env file. Will be used as fallback when combined with other options. }
-                {--flush : Delete all existing dumps in the dump folder when using a remote dump. }
                 {--l|latest : Import the most recent dump available in the configured dumps directory. }
                 {--m|migrate : Run database migrations after import. }
                 {--w|no-wipe : Do not wipe the database before importing the dump. }';
@@ -57,6 +56,8 @@ class ImportDump extends Command
     protected const string DOWNLOAD_REMOTE_DUMP = 'Download remote dump';
     protected const string IMPORT_EXISTING_LOCAL_DUMP = 'Import existing dump';
     protected Protector $protector;
+    protected Filesystem $sourceDisk;
+    protected bool $needsCleanup = false;
 
     /**
      * Execute the console command.
@@ -66,8 +67,6 @@ class ImportDump extends Command
      */
     public function handle(): int
     {
-        $this->newLine();
-
         if (App::environment('production') && !$this->option('allow-production')) {
             throw new InvalidEnvironmentException(
                 'Import is not allowed on production systems! Use --allow-production'
@@ -82,6 +81,7 @@ class ImportDump extends Command
 
         $this->protector = $protectorConfigurator->makeProtector();
         $this->protector->guardRequiredFunctionsEnabled();
+        $this->sourceDisk = $this->protector->getStorageDisk();
 
         $hasFile = !empty(trim($this->option('file')));
 
@@ -98,39 +98,25 @@ class ImportDump extends Command
             default => $this->getDumpInteractive(),
         };
 
-        $this->runImport($dumpSource, $this->option('force'));
-
-        $this->newLine();
+        $this->runImport($dumpSource);
 
         return self::SUCCESS;
     }
 
-    /**
-     * Reads the remote dump file and deletes all old dumps if the flush option is set.
-     */
-    protected function getDumpFromRemote(): array
+    protected function getDumpFromRemote(): string
     {
-        info(sprintf('Downloading dump from remote server to disk %s, path %s', $this->protector->getLocalDiskName(), $this->protector->getLocalDiskBaseDirectory()));
+        $dumpPath = $this->protector->createLocalFilePath();
+        $this->sourceDisk = $this->protector->getLocalDisk();
+        $this->needsCleanup = true;
 
-        $relativeRemoteDumpFilePath = $this->protector->createLocalFilePath();
-
-        $this->protector->download(
-            disk: $this->protector->getLocalDisk(),
-            filePath: $relativeRemoteDumpFilePath,
+        spin(
+            callback: fn() => $this->protector->download(disk: $this->sourceDisk, filePath: $dumpPath),
+            message: 'Downloading dump...'
         );
 
         info('Successfully retrieved remote dump.');
 
-        if ($this->option('flush')) {
-            $this->protector->flush();
-            warning(sprintf('Deleted all old files on disk %s in path %s', $this->protector->getStorageDiskName(), $this->protector->getStorageDiskBaseDirectory()));
-        }
-
-        return [
-            'path' => $relativeRemoteDumpFilePath,
-            'disk' => $this->protector->getLocalDisk(),
-            'cleanup' => true,
-        ];
+        return $dumpPath;
     }
 
     /**
@@ -139,49 +125,34 @@ class ImportDump extends Command
      *
      * @throws FileNotFoundException
      */
-    protected function getDumpFromFile(): array
+    protected function getDumpFromFile(): string
     {
         $isAbsoluteFilePath = $this->isAbsolutePath($this->option('file'));
 
-        switch ($isAbsoluteFilePath) {
-            case true:
-                $dumpFilePath = $this->option('file');
+        if ($isAbsoluteFilePath) {
+            $absoluteDumpPath = $this->option('file');
 
-                if (!file_exists($dumpFilePath)) {
-                    throw new FileNotFoundException($dumpFilePath);
-                }
+            if (!file_exists($absoluteDumpPath)) {
+                throw new FileNotFoundException($absoluteDumpPath);
+            }
 
-                return [
-                    'path' => $dumpFilePath,
-                    'disk' => null,
-                    'cleanup' => false,
-                ];
-            default:
-                // This will throw an exception if the dump file was not found.
-                $dumpFilePath = $this->protector->getDumpFile($this->option('file'));
-
-                return [
-                    'path' => $dumpFilePath,
-                    'disk' => null,
-                    'cleanup' => false,
-                ];
+            return $absoluteDumpPath;
         }
+
+        // This will throw an exception if the dump file was not found.
+        return $this->protector->dumpFile($this->option('file'));
     }
 
-    protected function getLatestDump(): array
+    protected function getLatestDump(): string
     {
-        $relativeImportFilePath = $this->protector->getLatestDumpName();
+        $dumpPath = $this->protector->getLatestDumpName();
 
-        info(sprintf('Importing %s', $relativeImportFilePath));
+        info(sprintf('Importing %s', $dumpPath));
 
-        return [
-            'path' => $relativeImportFilePath,
-            'disk' => null,
-            'cleanup' => false,
-        ];
+        return $dumpPath;
     }
 
-    protected function getDumpInteractive(): array
+    protected function getDumpInteractive(): string
     {
         if ($this->userWantsRemoteDump()) {
             return $this->getDumpFromRemote();
@@ -193,56 +164,48 @@ class ImportDump extends Command
     /**
      * Returns the file path to a selected dump.
      */
-    protected function chooseImportDump(?string $connectionName): array
+    protected function chooseImportDump(?string $connectionName): string
     {
         $connectionFiles = $this->getConnectionFiles($connectionName)->keys();
 
         if ($connectionFiles->count() === 1) {
-            $relativeImportFilePath = $connectionFiles->first();
+            $dumpPath = $connectionFiles->first();
 
-            info(sprintf('Using file "%s" because there are no other dumps.', $relativeImportFilePath));
-        } else {
-            $importFile = select(
-                label: 'Which file do you want to import?',
-                options: $connectionFiles->mapWithKeys(fn(string $file) => [$file => $file])->toArray(),
-            );
+            info(sprintf('Using file "%s" because there are no other dumps.', $dumpPath));
 
-            $relativeImportFilePath = $connectionFiles->firstWhere(fn($file) => $file === $importFile);
+            return $dumpPath;
         }
 
-        return [
-            'path' => $relativeImportFilePath,
-            'disk' => null,
-            'cleanup' => false,
-        ];
+        $selectedFile = select(
+            label: 'Which file do you want to import?',
+            options: $connectionFiles,
+        );
+
+        return $connectionFiles->firstWhere(fn($file) => $file === $selectedFile);
     }
 
     /**
      * Imports the selected SQL dump.
      */
-    protected function runImport(array $dumpSource, ?bool $optionForce): void
+    protected function runImport(string $dumpPath): void
     {
-        /** @var string $sourcePath */
-        $sourcePath = $dumpSource['path'];
-        /** @var ?Filesystem $sourceDisk */
-        $sourceDisk = $dumpSource['disk'];
-        /** @var bool $cleanup */
-        $cleanup = $dumpSource['cleanup'];
-
         try {
-            if ($optionForce || confirm(
+            if ($this->option('force') || confirm(
                     sprintf(
                         'Are you sure that you want to import the dump into the database: %s?',
                         $this->protector->getDatabaseName()
                     )
                 )) {
-                spin(fn() => $this->protector->import(
-                    $sourcePath,
-                    $sourceDisk,
-                    noWipe: $this->option('no-wipe'),
-                    migrate: $this->option('migrate'),
-                    allowProduction: $this->option('allow-production'),
-                ), 'Importing dump...');
+                spin(
+                    callback: fn() => $this->protector->import(
+                        $dumpPath,
+                        $this->sourceDisk,
+                        noWipe: $this->option('no-wipe'),
+                        migrate: $this->option('migrate'),
+                        allowProduction: $this->option('allow-production'),
+                    ),
+                    message: 'Importing dump...'
+                );
 
                 info('Import done!');
 
@@ -251,10 +214,11 @@ class ImportDump extends Command
 
             info('Import aborted');
         } finally {
-            if ($cleanup && !$this->isAbsolutePath($sourcePath)) {
-                ($sourceDisk ?? $this->protector->getStorageDisk())->delete([
-                    $sourcePath,
-                    $sourcePath . '.meta',
+            // Clean-up local in case there was a dump downloaded from remote.
+            if ($this->needsCleanup && !$this->isAbsolutePath($dumpPath)) {
+                $this->protector->getLocalDisk()->delete([
+                    $dumpPath,
+                    $dumpPath . '.meta',
                 ]);
             }
         }
@@ -267,7 +231,7 @@ class ImportDump extends Command
      */
     protected function getConnectionFiles(?string $connectionName = null): Collection
     {
-        $sortedFiles = $this->protector->getDumpFilesWithMetadata()
+        $sortedFiles = $this->protector->dumpFilesWithMetadata()
             ->sortByDesc(
             // Supporting the legacy format.
                 fn($file) => Arr::get($file, 'meta.database.dumpedAtDate') ?? Arr::get($file, 'meta.dumpedAtDate')
@@ -286,10 +250,6 @@ class ImportDump extends Command
 
         if ($sortedFiles->isEmpty()) {
             throw new EmptyBaseDirectoryException();
-        }
-
-        if ($this->option('ignore-connection-filter')) {
-            return $sortedFiles;
         }
 
         $filesByConnection = $sortedFiles->groupBy(
@@ -337,12 +297,11 @@ class ImportDump extends Command
     {
         if ($connectionNames->count() === 1) {
             $connectionName = $connectionNames->firstOrFail();
-            $connectionLabel = $this->getConnectionDisplayName($connectionName);
 
             info(
                 sprintf(
                     'Using connection "%s" because there are no dumps created through other connections.',
-                    $connectionLabel
+                    $connectionName
                 )
             );
 
@@ -351,20 +310,12 @@ class ImportDump extends Command
 
         return select(
             label: 'Import dump for which connection?',
-            options: $connectionNames->mapWithKeys(fn(string $name) => [$name => $this->getConnectionDisplayName($name)])->toArray(),
+            options: $connectionNames,
         );
-    }
-
-    protected function getConnectionDisplayName(string $connectionName): string
-    {
-        return $connectionName === static::UNKNOWN_CONNECTION_NAME
-            ? static::UNKNOWN_CONNECTION_LABEL
-            : $connectionName;
     }
 
     protected function isAbsolutePath(string $filePath): bool
     {
-        return str_starts_with($filePath, DIRECTORY_SEPARATOR)
-            || preg_match('/^[A-Za-z]:[\\\\\/]/', $filePath) === 1;
+        return Str::startsWith($filePath, DIRECTORY_SEPARATOR);
     }
 }